Refactor tradesToPublish test
